# docs/api/openapi.yaml
openapi: "3.1.0"

info:
  title: Spring Boot AI Template API
  description: |
    REST API built with **Java 21 & Spring Boot 3**.

    **API-first design**: this file is the single source of truth for the API contract.
    The Spring controllers in `src/main/java/com/example/template/api/v1/` implement this spec.
  version: "1.0.0"
  contact:
    name: API Support
    email: support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT

servers:
  - url: http://localhost:8080
    description: Local development (profile=local)

tags:
  - name: Users
    description: User lifecycle management
  - name: Health
    description: Actuator health & readiness probes

# ── Paths ─────────────────────────────────────────────────────────────────────
paths:

  /api/v1/users:
    post:
      tags: [Users]
      summary: Create a new user
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
            example:
              email: jane.doe@example.com
              firstName: Jane
              lastName: Doe
      responses:
        "201":
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        "400":
          $ref: '#/components/responses/ValidationError'
        "409":
          $ref: '#/components/responses/Conflict'

    get:
      tags: [Users]
      summary: List all users (paginated)
      operationId: listUsers
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
        - $ref: '#/components/parameters/Sort'
      responses:
        "200":
          description: Paginated user list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedUserResponse'

  /api/v1/users/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
        description: User UUID

    get:
      tags: [Users]
      summary: Get user by ID
      operationId: getUserById
      responses:
        "200":
          description: User found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        "404":
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Users]
      summary: Update user (patch semantics — null fields ignored)
      operationId: updateUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        "400":
          $ref: '#/components/responses/ValidationError'
        "404":
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Users]
      summary: Delete user
      operationId: deleteUser
      responses:
        "204":
          description: User deleted
        "404":
          $ref: '#/components/responses/NotFound'

  /actuator/health:
    get:
      tags: [Health]
      summary: Application health (liveness + readiness)
      operationId: health
      responses:
        "200":
          description: Application is healthy

# ── Components ────────────────────────────────────────────────────────────────
components:

  schemas:

    CreateUserRequest:
      type: object
      required: [email, firstName, lastName]
      properties:
        email:
          type: string
          format: email
          example: jane.doe@example.com
        firstName:
          type: string
          minLength: 1
          maxLength: 100
          example: Jane
        lastName:
          type: string
          minLength: 1
          maxLength: 100
          example: Doe

    UpdateUserRequest:
      type: object
      description: All fields are optional — only provided (non-null) fields are applied.
      properties:
        firstName:
          type: string
          minLength: 1
          maxLength: 100
        lastName:
          type: string
          minLength: 1
          maxLength: 100

    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
        email:
          type: string
          format: email
          example: jane.doe@example.com
        firstName:
          type: string
          example: Jane
        lastName:
          type: string
          example: Doe
        status:
          $ref: '#/components/schemas/UserStatus'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    PagedUserResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/UserResponse'
        page:
          type: integer
          example: 0
        size:
          type: integer
          example: 20
        totalElements:
          type: integer
          format: int64
          example: 100
        totalPages:
          type: integer
          example: 5
        last:
          type: boolean
          example: false

    UserStatus:
      type: string
      enum: [ACTIVE, INACTIVE, SUSPENDED]

    ErrorResponse:
      type: object
      properties:
        errorCode:
          type: string
          example: RESOURCE_NOT_FOUND
        message:
          type: string
          example: "User with id '550e8400-...' not found"
        path:
          type: string
          example: /api/v1/users/550e8400-e29b-41d4-a716-446655440000
        timestamp:
          type: string
          format: date-time
        fieldErrors:
          type: array
          items:
            $ref: '#/components/schemas/FieldError'

    FieldError:
      type: object
      properties:
        field:
          type: string
          example: email
        message:
          type: string
          example: Must be a valid email address

  parameters:

    Page:
      name: page
      in: query
      schema:
        type: integer
        default: 0
        minimum: 0
      description: Page number (0-based)

    Size:
      name: size
      in: query
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
      description: Page size

    Sort:
      name: sort
      in: query
      schema:
        type: string
        example: "createdAt,desc"
      description: "Sort expression: field,(asc|desc)"

  responses:

    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

    ValidationError:
      description: Request validation failed (400)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

    Conflict:
      description: Resource conflict (409)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'